+++ src/utils/utils_posix_common.c
@@ -17,7 +17,9 @@
 static UTIL_ONCE_FLAG Page_size_is_initialized = UTIL_ONCE_FLAG_INIT;
 static size_t Page_size;
 
-static void _util_get_page_size(void) { Page_size = sysconf(_SC_PAGE_SIZE); }
+static void _util_get_page_size(void) {
+    Page_size = (size_t)sysconf(_SC_PAGE_SIZE);
+}
 
 size_t util_get_page_size(void) {
     util_init_once(&Page_size_is_initialized, _util_get_page_size);
